roughly finishes restructuring the ml package
