+++ rr_common/CMakeLists.txt
@@ -55,5 +55,4 @@ add_subdirectory(src/planner)
 add_subdirectory(src/image_pcl_converter)
 add_subdirectory(src/mapper)
 add_subdirectory(src/navigation_controller)
-add_subdirectory(src/nn_planner)
 add_subdirectory(src/sound_node)
