--- rr_common/src/nn_planner/CMakeLists.txt
@@ -1,5 +1,0 @@
-catkin_install_python(
-        PROGRAMS
-        nn_planner.py
-        DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
-)
