--- rr_ml/end_to_end/example_set.py
@@ -1,32 +1,0 @@
-import cPickle as pickle
-import lz4
-import numpy as np
-
-
-class Example:
-    def __init__(self, img, angle):
-        self.img_buf = lz4.compress(img)
-        self.img_shape = img.shape
-        self.angle = angle
-
-    def get_image(self):
-        return np.frombuffer(lz4.decompress(self.img_buf),dtype='uint8').reshape(self.img_shape)
-
-class ExampleSet:
-    def __init__(self):
-        self.train = []
-        self.test = []
-
-    def add(self, other):
-        self.train += other.train
-        self.test += other.test
-
-    def save(self, fname):
-        with open(fname, 'w') as f:
-            f.write(lz4.compress(pickle.dumps(self)))
-
-    @staticmethod
-    def load(fname):
-        with open(fname, 'r') as f:
-            obj = pickle.loads(lz4.decompress(f.read()))
-        return obj
