+++ rr_ml/launch/convert_data_e2e.launch
@@ -1,0 +1,17 @@
+<launch>
+    <arg name="dataset"/>
+    <arg name="bag"/>
+    <arg name="bag2" default=""/>
+    <arg name="bag3" default=""/>
+    <arg name="bag4" default=""/>
+    <arg name="bag5" default=""/>
+
+    <node pkg="rr_ml" type="convert_data_e2e" name="convert_e2e" output="screen">
+        <param name="output_dir" value="$(find rr_ml)/exported_data/end_to_end/$(arg dataset)"/>
+        <param name="bag_file_0" value="$(arg bag)"/>
+        <param name="bag_file_1" value="$(arg bag2)"/>
+        <param name="bag_file_2" value="$(arg bag3)"/>
+        <param name="bag_file_3" value="$(arg bag4)"/>
+        <param name="bag_file_4" value="$(arg bag5)"/>
+    </node>
+</launch>
