+++ rr_ml/nodes/CMakeLists.txt
@@ -1,0 +1,5 @@
+catkin_install_python(PROGRAMS
+        run_e2e
+        train_e2e
+        convert_data_e2e
+        DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
