+++ rr_ml/nodes/convert_data_e2e
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+from end_to_end import convert
+
+if __name__ == '__main__':
+    convert.main()
