+++ rr_ml/nodes/run_e2e
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+from end_to_end import run
+
+if __name__ == '__main__':
+    run.main()
