+++ rr_ml/nodes/train_e2e
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+from end_to_end import train
+
+if __name__ == '__main__':
+    train.main()
